Relay for TXS

+++ Makefile
@@ -42,9 +42,10 @@ deps:
 	. ./util/setup.sh
 
 bins:
-	#Build and install genesis tool, libra-node, and miner
+# Build and install genesis tool, libra-node, and miner
 	cargo run -p stdlib --release
-	# NOTE: stdlib is built for cli bindings
+
+# NOTE: stdlib is built for cli bindings
 	cargo build -p libra-node -p miner -p backup-cli -p ol-cli -p txs --release
 
 install:
@@ -142,11 +143,11 @@ reg:
 	--shared-backend ${REMOTE}
 	
 
-## Helpers to verify the local state.
+# Helpers to verify the local state.
 verify:
 	cargo run -p libra-genesis-tool --release --  verify \
 	--validator-backend ${LOCAL}
-	# --genesis-path ${DATA_PATH}/genesis.blob
+# --genesis-path ${DATA_PATH}/genesis.blob
 
 verify-gen:
 	cargo run -p libra-genesis-tool --release --  verify \
@@ -155,11 +156,11 @@ verify-gen:
 
 
 #### GENESIS  ####
-build-gen:
-	cargo run -p libra-genesis-tool --release -- genesis \
-	--chain-id ${CHAIN_ID} \
-	--shared-backend ${REMOTE} \
-	--path ${DATA_PATH}/genesis.blob
+# build-gen:
+# 	cargo run -p libra-genesis-tool --release -- genesis \
+# 	--chain-id ${CHAIN_ID} \
+# 	--shared-backend ${REMOTE} \
+# 	--path ${DATA_PATH}/genesis.blob
 
 genesis:
 	cargo run -p libra-genesis-tool --release -- files \
@@ -175,6 +176,10 @@ genesis:
 start:
 # run in foreground. Only for testing, use a daemon for net.
 	cargo run -p libra-node -- --config ${DATA_PATH}/validator.node.yaml
+
+# Start a fullnode instead of a validator node
+start-full:
+	cargo run -p libra-node -- --config ${DATA_PATH}/fullnode.node.yaml
 
 daemon:
 # your node's custom libra-node.service lives in ~/.0L. Take the template from libra/util and edit for your needs.
@@ -200,12 +205,16 @@ daemon:
 #### TEST SETUP ####
 
 clear:
-	if test ${DATA_PATH}/key_store.json; then \
+ifeq (${TEST}, y)
+
+	@if test -d ${DATA_PATH}; then \
 		cd ${DATA_PATH} && rm -rf libradb *.yaml *.blob *.json db *.toml; \
 	fi
-	if test -d ${DATA_PATH}/blocks; then \
+	@if test -d ${DATA_PATH}/blocks; then \
 		rm -f ${DATA_PATH}/blocks/*.json; \
 	fi
+endif
+
 
 fixture-stdlib:
 	make stdlib
@@ -228,9 +237,10 @@ check:
 
 fix:
 ifdef TEST
-	echo ${NS}
-	@if test ! -d ${0L_PATH}; then \
-		mkdir ${0L_PATH}; \
+	@echo NAMESPACE: ${NS}
+	@echo GENESIS: ${V}
+	@if test ! -d ${DATA_PATH}; then \
+		echo Creating Directories \
 		mkdir ${DATA_PATH}; \
 		mkdir -p ${DATA_PATH}/blocks/; \
 	fi
@@ -259,12 +269,10 @@ endif
 #### HELPERS ####
 set-waypoint:
 	@if test -f ${DATA_PATH}/key_store.json; then \
-		jq -r '. | with_entries(select(.key|match("/waypoint";"i")))[].value' ${DATA_PATH}/key_store.json > ${DATA_PATH}/client_waypoint; \
+		jq -r '. | with_entries(select(.key|match("-oper/waypoint";"i")))[].value' ${DATA_PATH}/key_store.json > ${DATA_PATH}/client_waypoint; \
+		jq -r '. | with_entries(select(.key|match("-oper/genesis-waypoint";"i")))[].value' ${DATA_PATH}/key_store.json > ${DATA_PATH}/genesis_waypoint; \
 	fi
 
-	@if test ! -f ${DATA_PATH}/key_store.json; then \
-		cat ${DATA_PATH}/restore_waypoint > ${DATA_PATH}/client_waypoint; \
-	fi
 	@echo client_waypoint:
 	@cat ${DATA_PATH}/client_waypoint
 
@@ -306,54 +314,59 @@ stop:
 ##### DEVNET TESTS #####
 # Quickly start a devnet with fixture files. To do a full devnet setup see 'devnet-reset' below
 
-devnet: 
-	MNEM='${MNEM}' make stop clear fix devnet-keys devnet-yaml start
+frozen: 
+# A QUICK TEST FROM FIXTURES. ASSUMES EVERYTHING WAS SETUP: new genesis-blobs, and mock archive infrastructure. For this see `devnet-archive` below 
+
 # runs a smoke test from fixtures. Uses genesis blob from fixtures, assumes 3 validators, and test settings.
-# This will work for validator nodes alice, bob, carol, and any fullnodes; 'eve'
 
-devnet-keys: 
-	MNEM='${MNEM}' cargo run -p miner -- init --skip-miner
+# This will work for validator nodes alice, bob, carol. New onboarded "eve" needs to run devnet-onboard
 
-devnet-yaml:
-	cargo run -p miner -- files
+	MNEM='${MNEM}' make stop clear fix dev-wizard start
 
-devnet-onboard: clear fix
-	#starts config for a new miner "eve", uses the devnet github repo for ceremony
-	MNEM='${MNEM}' cargo r -p miner -- init --skip-miner
-	cargo r -p miner -- files
+dev-wizard:
+# starts config for a new miner "eve", uses the devnet github repo for ceremony
+# get genesis.blcok from MOCK genesis store OLSF/dev-genesis
+	MNEM='${MNEM}' cargo run -p miner -- val-wizard --skip-mining --skip-fetch-genesis --chain-id 1 --github-org OLSF --repo dev-genesis
 
-### FULL DEVNET RESET ####
 
-devnet-reset: 
-	MNEM='${MNEM}' make devnet-reset-ceremony genesis start
-# Tests the full genesis ceremony cycle, and rebuilds all genesis and waypoints.
+dev-join: clear fix dev-wizard
+# REQUIRES MOCK GIT INFRASTRUCTURE: OLSF/dev-genesis OLSF/dev-epoch-archive
+# see `devnet-archive` below 
+# We want to simulate the onboarding/new validator fetching genesis files from the mock archive: dev-genesis-archive
 
-devnet-reset-ceremony:
-# note: this uses the NS in local env to create files i.e. alice or bob
-# as a operator/owner pair.
-	SKIP_BLOB=y make clear fix
+# mock restore backups from dev-epoch-archive
+	rm -rf ~/.0L/restore
+# restore from MOCK archive OLSF/dev-epoch-archive
+	cargo r -p ol-cli -- restore
+# start a node with fullnode.node.yaml configs
+	make start-full
+
+### FULL DEVNET E2E ####
+
+devnet:
+	MNEM='${MNEM}' make genesis start
+
+dev-register: clear fix
 	echo ${MNEM} | head -c -1 | make register
 
-devnet-reset-onboard: clear 
-# fixtures needs a file that works
-	SKIP_BLOB=y make fix
-	cp ${SOURCE}/fixtures/genesis/${V}/* ~/.0L/
-# # starts config for a new miner "eve", uses the devnet github repo for ceremony
-	MNEM="${MNEM}" cargo r -p miner -- val-wizard  --skip-fetch-genesis --chain-id 1 --github-org OLSF --repo dev-genesis --skip-mining 
+# devnet-setup-onboard: clear 
+# # fixtures needs a file that works
+# 	SKIP_BLOB=y make fix
+# 	cp ${SOURCE}/fixtures/genesis/${V}/* ~/.0L/
+# # # starts config for a new miner "eve", uses the devnet github repo for ceremony
+# 	MNEM="${MNEM}" cargo r -p miner -- val-wizard  --skip-fetch-genesis --chain-id 1 --github-org OLSF --repo dev-genesis --skip-mining 
 
-#### GIT HELPERS FOR DEVNET AUTOMATION ####
-devnet-save-genesis: set-waypoint
+#### PERSIST THE MOCK ARCHIVES TO DEVNET INFRASTRUCTURE ####
+
+# usually do this on Alice, which has the dev-epoch-archive repo, and dev-genesis
+dev-infra: dev-save-genesis dev-backup-archive
+
+dev-save-genesis: set-waypoint
 	rsync -a ${DATA_PATH}/genesis* ${SOURCE}/fixtures/genesis/${V}/
 	git add ${SOURCE}/fixtures/genesis/${V}/
-	git commit -a -m "save genesis fixtures to ${V}"
-	git push
+	git commit -a -m "save genesis fixtures to ${V}" | true
+	git push | true
 
-devnet-hard:
-	git reset --hard origin/${V} 
+dev-backup-archive:
+	cd ${HOME}/dev-epoch-archive && make devnet-backup
 
-devnet-pull:
-# must be on a branch
-	git fetch && git checkout ${V} -f && git pull
-
-devnet-fullnode:
-	cargo run -p miner -- fn-wizard --path ~/.0L/
